iot-locker/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.1
PyMySQL==1.1.0
marshmallow==3.21.1
marshmallow-sqlalchemy==0.29.0
marshmallow-enum==1.5.1
passlib==1.7.4
cryptography==42.0.5